made default handler more generic
Should we split `PaymentsDb` into multiple databases? At least at the
interface level.
